move health stuff to its own package for cleaner code
