+++ gateway-api/src/main/java/io/shipkit/gatewayapi/gatewayapi/core/security/account/AccountGraphQLController.java
@@ -2,12 +2,10 @@ package io.shipkit.gatewayapi.gatewayapi.core.security.account;
 
 import io.shipkit.gatewayapi.gatewayapi.core.security.account.dto.AuthPayloadDTO;
 import io.shipkit.gatewayapi.gatewayapi.core.security.account.dto.CreateAccountDTO;
-import io.shipkit.gatewayapi.gatewayapi.core.security.account.dto.StatusDTO;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import org.springframework.graphql.data.method.annotation.Argument;
 import org.springframework.graphql.data.method.annotation.MutationMapping;
-import org.springframework.graphql.data.method.annotation.QueryMapping;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 
@@ -27,11 +25,5 @@ public class AccountGraphQLController {
     @PreAuthorize("permitAll()")
     public AuthPayloadDTO login(@Argument String email, @Argument String password) {
         return accountService.login(email, password);
-    }
-
-    @QueryMapping
-    @PreAuthorize("permitAll()")
-    public StatusDTO status() {
-        return new StatusDTO("healthy", accountService.isAdminInitialized());
     }
 }
