+++ gateway-api/src/main/resources/graphql/health.graphqls
@@ -1,0 +1,8 @@
+type Status {
+    status: String!
+    adminInitialized: Boolean!
+}
+
+extend type Query {
+    status: Status!
+}
